+++ src/main/java/com/smarthouse/user/service/RoleServiceImpl.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
 
 @Service
-public class RoleServiceImpl extends ServiceImpl<Role, Integer> {
+public class RoleServiceImpl extends ServiceImpl<Role, Long> {
     @Autowired
-    public RoleServiceImpl(final JpaRepository<Role, Integer> repository) {
+    public RoleServiceImpl(final JpaRepository<Role, Long> repository) {
         super(repository);
     }
 }
